+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Bump: sentry-android to v4.0.0 #1309
 - build(ios): Bump sentry-cocoa to 6.1.4 #1308
 - fix: Handle auto session tracking start on iOS #1308
+- feat: Use beforeNavigate in routing instrumentation to match behavior on JS #1313
 
 ## 2.2.0-beta.0